do guest driver to host memcpys using sse.
Just copy chunks of data in sse2 instrcutions with prefeching of cache
bytes. - 35%~ cpu improvment when playing video in windows.

The first code that I wrote for this patch actuly handled
all the algiment requirment in the spice qxl win driver,
and did all the copys in 16 bytes algiment,
however due requiments from libspice server the cost of maitaing 16byte algiment
for each "line of copy" is too high and therefore whenever we better use
unalgiment copys whenever the bitmaps arent already 16byte algiment.

There is alot of space for improvment in this patch, but due to
timelines and the fact that we have another hurry tasks I will just push
this patch for now - the results are very good right now

Thanks.

Signed-off-by: Izik Eidus <ieidus@redhat.com>