Move SurfaceInfo to global data
When we release a surface resource, for instance when AllocMem gets
OOM, we call ReleaseOutput with the pdev in use (i.e. probably the
currently active one). However, its possible that the resource we
free is actually from another (now inactive) pdev.

Since the surface release function FreeDelSurface() uses pdev to
find the data for the surface id we may find the wrong surface data.

So, we move all the SurfaceInfos except the primary one to a global
array. This is fine since the surface ids (except 0) are never shared
between pdevs due to the surfaces_used global array.