Release resources before allocating
Before trying to allocate memory we free a large chunk of
freeable resources. Leaving the resources unfreed before allocating
increases the risks of unnecessarily fragmenting the heap, which
is bad, especially for the vram which needs large free contiguous
allocations for surfaces.

We do somewhat limit the number of resources we free on each call
to ensure that we get an approximately even spread of the cost of
allocations.