Make global_res be an array of pointers and Res a pointer
Instead of allocating an array of DevRes we allocate each DevRes
dynamically for each element.

Also make PDEV->Res be a pointer instead of copying the global pdev.
This also drops the needto Sync it when the pdev is made inactive,
as all pdevs share the same DevRes.

This way shared things like semaphores are really shared.