driver: allowed modes: remove limit of 480 <= {X,Y} <= 2000
Later todo is to check that there is available video memory for
this, but that requires either access to PDev or probably dynamically
checking anyway to make sure we actually fit. Another option is
to leave that check in the host (when we do a IO_SETMODE)

Bumping version to 1.4.1.1

Tested with winxp 32 bit only.