display/res: fix ie9 rendering
Internet Explorer 9 renders cursors and other things by reusing the same
SURFOBJ (i.e. surfobj->iUniq is constant) but changing the pallete.
We were wrongly ignoring the pallette's iUniq because the XO_TABLE flag
was not set. That flag should not be checked when calculating the key for
our cache. When that flag is ignored we correctly calculate a key that uses
both the surf->iUniq and the colortrans->iUniq together (64 bit from two 32
bit values).