display/surface: CreateDeviceBitmap: remove redundant FreeSurface
on error2 path, if we failed QXLGetSurface, we free the surface info but then
the caller (DrvCreateDeviceBitmap) frees it again. In addition, we cannot free
the SurfaceInfo since it is the handle given to GDI and is accessed on callback
from EngDestroySurface (which is called immediatelly after in the same error
path).