display/*: add PDev->enabled
GDI will continue using any callback we registered even after a
DrvAssertMode(FALSE). We are expected to move any surface we own to GDI handled
and ignore any new requests to create a surface. This is called punting and we
use PDev->enabled to indicate if this is required.

A later patch will set PDev->enabled to FALSE on DrvAssertMode.