asynchronous io port support (introduced in revision 3 of qxl device)
Fixes same issue in RHBZ#700134, but for a windows guest. Requires a revision 3
pci device, that will be introduced with qemu patches. If the revision is 2 the
old behavior is maintained, namely using the non asynchronous io ports.

qxl revision 3 (QXL_REVISION_V10) gains support for async io operations for

 UPDATE_AREA
 CREATE_PRIMARY
 DESTROY_PRIMARY
 DESTROY_SURFACE_WAIT (not used currently, just exported)
 DESTROY_SURFACES

use that if the revision is 3 or higher.

Async io ports let qemu complete the io without blocking on spice, and issue an
interrupt when the operation started by the io write is complete. This also means
less time when the qemu global mutex is held.