display: not using globals, RHBZ #731644
due to the previous patch ("support clearing the pci-bar when the pdev is disabled for revision<3 as well"),
surfaces_info and other resources like caches, are no longer shared between pdevs. Moreover,
it is an error to use globals, since different display driver sessions have different memory space.

In addition I removed all the semaphores that are not related to the pci ram. These semaphores
were protecting the surfaces_info array and the caches, however these sturctures belong only to one pdev,
and according to http://msdn.microsoft.com/en-us/library/ff568281%28v=vs.85%29.aspx, can be accessed
only from a single thread.

For the rest of the semaphores, I returned them to the same state that was in revision 1 => they
are part of the PDev. However, this should be fixed:
1) In order to protect the device log buffer,
   the print_sem must be shared by different pdevs and
   different display sessions. Alternatively, we can use IOCTL for logging.
2) malloc_sem: not sure what it protects. Maybe globals in mspace?
   since only the enabled pdev is allocating memory, I don't
   think it is required (unless it is possible to have
   AssertMode(x, enable) before AssertMode(y, disable).
3) cmd_sem, cursor_sem: again, since only the enabled pdev touches the cmd rings
   I don't think it is required.
4) io_sem - same as print sem. Note that we should prevent starvation between
            print_sem and io_sem in DebugPrintV.