display: handle interrupt handler disappearance (rhbz 721118)
This patch handles the surprise removal whql test. The surprise
removal test sends a surprise removal IRP via a filter driver to
qxl.sys. The handling of that IRP is actually done in videoprt.sys,
and there is no API to get notified. The side effect of the handler
in videoprt.sys is disabling our interrupt handler. This patch takes
care of continuing to work after our interrupt is disabled. This is
in effect the opposite of what the MSDN states that a device should do
after a surprise removal, since the MSDN requires every IO to fail after
the surprise removal IRP has been successfully handled. On the other
hand, this workaround is sufficient to pass the test.

The workaround is to turn the EngWaitForSingleEvent on the io_cmd_event
event into a timed wait.

If it fails, we revert to synchronouse IO, which doesn't require that
event at all.