miniport: QXLEscape: use 5/5/5 16 bit
Change IOCTL_QXL_SET_CUSTOM_DISPLAY to use 5/6/5 (via SetCustomDisplay).

The replaced 5/6/5 is not supported by us. Specifically we fail
display/rop.c:TestSrcBits and thus fail _BitBlt. I'm not sure how
exactly this causes the BSOD for 863410, but it does in fact do that,
and applying this fix, by not failing _BitBlt also fixes 863410.

RHBZ: 863410