only allow word aligned strides
This is a pixman requirement. If it isn't enforced spice crashes, qemu
with it. That should be fixed, but there is no simple way to allow for
non word aligned strides, and the only down side is that for 16 bit
modes we will require even widths. And we already have the same
implementation detail in the X driver, only there it is a multiple of 8,
not 4.

The SurfaceCreate command's stride is computed in one place:
 miniport:FillVidModeInfo sets ScreenStride as word aligned
 display:pdev->stride is set from ScreenStride

A third change at GetSurfaceMemory fixes the stride used for GDI
allocated surface, but since surface 0 has a fixed larger then required
area this isn't required for the bug fix, just for consistency.